Fixed other api inconsistency.

+++ toxcore/Messenger.c
@@ -659,7 +659,7 @@ int m_set_usertyping(Messenger *m, int32_t friendnumber, uint8_t is_typing)
     return 0;
 }
 
-int m_get_istyping(Messenger *m, int32_t friendnumber)
+uint8_t m_get_istyping(Messenger *m, int32_t friendnumber)
 {
     if (friend_not_valid(m, friendnumber))
         return -1;
