+++ toxcore/tox.c
@@ -316,7 +316,7 @@ int tox_set_user_is_typing(Tox *tox, int32_t friendnumber, uint8_t is_typing)
  * returns 0 if friend is not typing.
  * returns 1 if friend is typing.
  */
-int tox_get_is_typing(Tox *tox, int32_t friendnumber)
+uint8_t tox_get_is_typing(Tox *tox, int32_t friendnumber)
 {
     Messenger *m = tox;
     return m_get_istyping(m, friendnumber);
